Got the tests running again
