+++ jest.config.js
@@ -1,27 +1,37 @@
 const { pathsToModuleNameMapper } = require("ts-jest");
-
 const { compilerOptions } = require("./tsconfig");
 
-/** @type {import('jest').Config} */
+const tsPreset = require("ts-jest/jest-preset");
+const angularPreset = require("jest-preset-angular/jest-preset");
+const { defaultTransformerOptions } = require("jest-preset-angular/presets");
+
+/** @type {import('ts-jest').JestConfigWithTsJest} */
 module.exports = {
-  reporters: ["default", "jest-junit"],
+  // ...tsPreset,
+  // ...angularPreset,
+  preset: "jest-preset-angular",
 
-  collectCoverage: true,
-  coverageReporters: ["html", "lcov"],
-  coverageDirectory: "coverage",
+  testEnvironment: "jsdom",
+  testMatch: ["**/+(*.)+(spec).+(ts)"],
 
-  moduleNameMapper: pathsToModuleNameMapper(compilerOptions?.paths || {}, {
-    prefix: "<rootDir>/",
-  }),
-  projects: [
-    "<rootDir>/jslib/angular/jest.config.js",
-    "<rootDir>/jslib/common/jest.config.js",
-    "<rootDir>/jslib/electron/jest.config.js",
-    "<rootDir>/jslib/node/jest.config.js",
-  ],
-
-  // Workaround for a memory leak that crashes tests in CI:
-  // https://github.com/facebook/jest/issues/9430#issuecomment-1149882002
-  // Also anecdotally improves performance when run locally
+  roots: ["<rootDir>"],
+  modulePaths: [compilerOptions.baseUrl],
+  moduleNameMapper: pathsToModuleNameMapper(compilerOptions.paths, { prefix: "<rootDir>/" }),
+  setupFilesAfterEnv: ["<rootDir>/test.setup.ts"],
   maxWorkers: 3,
+
+  transform: {
+    "^.+\\.tsx?$": [
+      "jest-preset-angular",
+      // 'ts-jest',
+      {
+        ...defaultTransformerOptions,
+        isolatedModules: true,
+        tsconfig: "./tsconfig.json",
+        astTransformers: {
+          before: ["./es2020-transformer.ts"],
+        },
+      },
+    ],
+  },
 };
