--- jslib/angular/jest.config.js
@@ -1,16 +1,0 @@
-const { pathsToModuleNameMapper } = require("ts-jest");
-
-const { compilerOptions } = require("../shared/tsconfig.libs");
-
-const sharedConfig = require("../shared/jest.config.angular");
-
-/** @type {import('jest').Config} */
-module.exports = {
-  ...sharedConfig,
-  displayName: "libs/angular tests",
-  preset: "jest-preset-angular",
-  setupFilesAfterEnv: ["<rootDir>/test.setup.ts"],
-  moduleNameMapper: pathsToModuleNameMapper(compilerOptions?.paths || {}, {
-    prefix: "<rootDir>/",
-  }),
-};
