--- jslib/common/test.setup.ts
@@ -1,17 +1,0 @@
-import { webcrypto } from "crypto";
-
-import { toEqualBuffer } from "./spec";
-
-Object.defineProperty(window, "crypto", {
-  value: webcrypto,
-});
-
-// Add custom matchers
-
-expect.extend({
-  toEqualBuffer: toEqualBuffer,
-});
-
-export interface CustomMatchers<R = unknown> {
-  toEqualBuffer(expected: Uint8Array | ArrayBuffer): R;
-}
