--- jslib/shared/tsconfig.libs.json
@@ -1,12 +1,0 @@
-{
-  "extends": "./tsconfig",
-  "compilerOptions": {
-    "paths": {
-      "tldjs": ["../common/src/misc/tldjs.noop"],
-      "jslib-common/*": ["../common/src/*"],
-      "jslib-angular/*": ["../angular/src/*"],
-      "jslib-electron/*": ["../electron/src/*"],
-      "jslib-node/*": ["../node/src/*"]
-    }
-  }
-}
