+++ test.setup.ts
@@ -1,5 +1,7 @@
 import { webcrypto } from "crypto";
+
 import "jest-preset-angular/setup-jest";
+import { toEqualBuffer } from "jslib/common/spec";
 
 Object.defineProperty(window, "CSS", { value: null });
 Object.defineProperty(window, "getComputedStyle", {
@@ -26,3 +28,15 @@ Object.defineProperty(document.body.style, "transform", {
 Object.defineProperty(window, "crypto", {
   value: webcrypto,
 });
+
+Object.defineProperty(window, "crypto", {
+  value: webcrypto,
+});
+
+expect.extend({
+  toEqualBuffer: toEqualBuffer,
+});
+
+export interface CustomMatchers<R = unknown> {
+  toEqualBuffer(expected: Uint8Array | ArrayBuffer): R;
+}
