+++ tsconfig.eslint.json
@@ -24,5 +24,5 @@
       }
     ]
   },
-  "include": ["src", "jslib", "scripts"]
+  "include": ["src", "jslib", "scripts", "./**.ts"]
 }
